# src/models/customer.py
from __future__ import annotations

from datetime import datetime, timezone
from typing import Optional

from sqlalchemy import DateTime, Index, Integer, String, Text
from sqlalchemy.orm import Mapped, mapped_column

from src.db.base_class import Base


def _utc_now() -> datetime:
    return datetime.now(timezone.utc)


class Customer(Base):
    """Member profile. vipTierName is derived on read; total_spent is checkout-owned after create."""

    id: Mapped[str] = mapped_column(String(64), primary_key=True)
    name: Mapped[str] = mapped_column(String(256), nullable=False)
    phone: Mapped[Optional[str]] = mapped_column(String(64), nullable=True, unique=True)
    email: Mapped[Optional[str]] = mapped_column(String(256), nullable=True, unique=True)
    vip_tier: Mapped[str] = mapped_column(String(32), nullable=False, default="regular")
    total_spent: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
    note: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
    created_at: Mapped[datetime] = mapped_column(
        DateTime(timezone=True), nullable=False, default=_utc_now
    )
    updated_at: Mapped[datetime] = mapped_column(
        DateTime(timezone=True), nullable=False, default=_utc_now, onupdate=_utc_now
    )

    __table_args__ = (
        Index("ix_customer_name", "name"),
        Index("ix_customer_vip_tier", "vip_tier"),
    )